+++ test/activation.test.ts
@@ -56,6 +56,7 @@ function makeVscode(): any {
       onDidGrantWorkspaceTrust: () => disposable(() => {}),
       onDidSaveTextDocument: () => disposable(() => {}),
       onDidCloseTextDocument: () => disposable(() => {}),
+      onDidRenameFiles: () => disposable(() => {}),
       onDidChangeTextDocument: () => disposable(() => {}),
       textDocuments: [],
     },